led off improvement

+++ led.h
@@ -1,4 +1,4 @@
-// Helper functions to control Brainlink's full color LED, more in LED.c 
+// Helper functions to control Brainlink's full color LED, more in LED.c
 
 // Set the pins and port of the LED
 #define REDLED    PIN0_bm
@@ -7,4 +7,4 @@
 #define LEDPORT   PORTD
 
 void init_led(void);
-void set_led(int red, int green, int blue);
+void set_led(uint8_t red, uint8_t green, uint8_t blue);
